# swagger.yaml
...
swagger: "2.0"
info:
  description: "The inner API of Popcube aim to define every methods need to manage\
    \ communication organisation inside PopCube chat project"
  version: "alpha"
  title: "Popcube iner API"
  contact:
    name: "Titouan Freville"
    email: "titouanfreville@gmail.com"
host: "api.popcube"
basePath: "/alpha"
schemes:
- "http"
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /folder:
    get:
      tags:
      - "Folders"
      summary: "Get folders"
      description: "This will get all the folders available in the organisation."
      operationId: "getAllFolder"
      parameters: []
      responses:
        200:
          description: "folderSlice Array of folders"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Folder"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
    post:
      tags:
      - "Folders"
      summary: "New folder"
      description: "This will create an folder for organisation folders library."
      operationId: "newFolder"
      parameters:
      - in: "body"
        name: "folder"
        description: "Link of the folder in server."
        required: false
        schema:
          $ref: "#/definitions/Folder"
        x-go-name: "Folder"
      responses:
        200:
          description: "folderObjectSuccess list of folders"
          schema:
            $ref: "#/definitions/Folder"
        422:
          description: "wrongEntityError is an error object to inform that the provided\
            \ object was not correctly formated"
          schema:
            $ref: "#/definitions/genericErrorModel"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
  /folder/all:
    get:
      tags:
      - "Folders"
      summary: "Get folders"
      description: "This will get all the folders available in the organisation."
      operationId: "getAllFolder1"
      parameters: []
      responses:
        200:
          description: "folderSlice Array of folders"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Folder"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
  /folder/link/{folderLink}:
    get:
      tags:
      - "Folders"
      summary: "Get folder from link"
      description: "This will return the folder object corresponding to provided link"
      operationId: "getFolderFromLink"
      parameters:
      - name: "folderLink"
        in: "path"
        description: "Link of the folder in server."
        required: true
        type: "string"
        x-go-name: "FolderLink"
      responses:
        200:
          description: "folderObjectSuccess list of folders"
          schema:
            $ref: "#/definitions/Folder"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
  /folder/message:
    post:
      tags:
      - "Folders"
      summary: "Get folders linked to message"
      description: "Return folders linked to provided message."
      operationId: "getFolderFromMessage"
      parameters:
      - in: "body"
        name: "message"
        description: "Link of the message in server."
        required: false
        schema:
          $ref: "#/definitions/Message"
        x-go-name: "Message"
      responses:
        200:
          description: "folderSlice Array of folders"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Folder"
        422:
          description: "wrongEntityError is an error object to inform that the provided\
            \ object was not correctly formated"
          schema:
            $ref: "#/definitions/genericErrorModel"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
  /folder/name/{folderName}:
    get:
      tags:
      - "Folders"
      summary: "Get folder from name"
      description: "This will return the folder object corresponding to provided name"
      operationId: "getFolderFromName"
      parameters:
      - name: "folderName"
        in: "path"
        description: "Link of the folder in server."
        required: true
        type: "string"
        x-go-name: "FolderName"
      responses:
        200:
          description: "folderObjectSuccess list of folders"
          schema:
            $ref: "#/definitions/Folder"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
  /folder/new:
    post:
      tags:
      - "Folders"
      summary: "New folder"
      description: "This will create an folder for organisation folders library."
      operationId: "newFolder1"
      parameters:
      - in: "body"
        name: "folder"
        description: "Link of the folder in server."
        required: false
        schema:
          $ref: "#/definitions/Folder"
        x-go-name: "Folder"
      responses:
        200:
          description: "folderObjectSuccess list of folders"
          schema:
            $ref: "#/definitions/Folder"
        422:
          description: "wrongEntityError is an error object to inform that the provided\
            \ object was not correctly formated"
          schema:
            $ref: "#/definitions/genericErrorModel"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
  /folder/type/{folderType}:
    get:
      tags:
      - "Folders"
      summary: "Get folder from type"
      description: "This will return the folder object corresponding to provided type"
      operationId: "getFolderFromType"
      parameters:
      - name: "folderType"
        in: "path"
        description: "Link of the folder in server."
        required: true
        type: "string"
        x-go-name: "FolderType"
      responses:
        200:
          description: "folderObjectSuccess list of folders"
          schema:
            $ref: "#/definitions/Folder"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
  /folder/{folderID}:
    put:
      tags:
      - "Folders"
      summary: "Update folder"
      description: "This will return the new folder object"
      operationId: "updateFolder"
      parameters:
      - name: "folderID"
        in: "path"
        description: "Link of the folder in server."
        required: true
        type: "integer"
        format: "int64"
        x-go-name: "FolderID"
      - in: "body"
        name: "folder"
        description: "Link of the folder in server."
        required: false
        schema:
          $ref: "#/definitions/Folder"
        x-go-name: "Folder"
      responses:
        200:
          description: "avatarObjectSuccess list of avatars"
          schema:
            $ref: "#/definitions/Avatar"
        422:
          description: "wrongEntityError is an error object to inform that the provided\
            \ object was not correctly formated"
          schema:
            $ref: "#/definitions/genericErrorModel"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
    delete:
      tags:
      - "Folders"
      summary: "Delete folder"
      description: "This will return an object describing the deletion"
      operationId: "deleteFolder"
      parameters:
      - name: "folderID"
        in: "path"
        description: "Link of the folder in server."
        required: true
        type: "integer"
        format: "int64"
        x-go-name: "FolderID"
      responses:
        200:
          description: "deleteMessage"
          schema:
            $ref: "#/definitions/deleteMessage"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
...
  /message:
    get:
      tags:
      - "Messages"
      summary: "Get messages"
      description: "This will get all the messages available in the organisation."
      operationId: "getAllMessage"
      parameters: []
      responses:
        200:
          description: "messageSlice Array of messages"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Message"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
    post:
      tags:
      - "Messages"
      summary: "New message"
      description: "This will create an message for organisation messages library."
      operationId: "newMessage"
      parameters:
      - in: "body"
        name: "message"
        description: "Link of the message in server."
        required: false
        schema:
          $ref: "#/definitions/Message"
        x-go-name: "Message"
      responses:
        200:
          description: "messageObjectSuccess list of messages"
          schema:
            $ref: "#/definitions/Message"
        422:
          description: "wrongEntityError is an error object to inform that the provided\
            \ object was not correctly formated"
          schema:
            $ref: "#/definitions/genericErrorModel"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
  /message/all:
    get:
      tags:
      - "Messages"
      summary: "Get messages"
      description: "This will get all the messages available in the organisation."
      operationId: "getAllMessage1"
      parameters: []
      responses:
        200:
          description: "messageSlice Array of messages"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Message"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
...
  /message/new:
    post:
      tags:
      - "Messages"
      summary: "New message"
      description: "This will create an message for organisation messages library."
      operationId: "newMessage1"
      parameters:
      - in: "body"
        name: "message"
        description: "Link of the message in server."
        required: false
        schema:
          $ref: "#/definitions/Message"
        x-go-name: "Message"
      responses:
        200:
          description: "messageObjectSuccess list of messages"
          schema:
            $ref: "#/definitions/Message"
        422:
          description: "wrongEntityError is an error object to inform that the provided\
            \ object was not correctly formated"
          schema:
            $ref: "#/definitions/genericErrorModel"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
  /message/{messageID}:
    put:
      tags:
      - "Messages"
      summary: "Update message"
      description: "This will return the new message object"
      operationId: "updateMessage"
      parameters:
      - name: "messageID"
        in: "path"
        description: "Link of the message in server."
        required: true
        type: "integer"
        format: "int64"
        x-go-name: "MessageID"
      - in: "body"
        name: "message"
        description: "Link of the message in server."
        required: false
        schema:
          $ref: "#/definitions/Message"
        x-go-name: "Message"
      responses:
        200:
          description: "messageObjectSuccess list of messages"
          schema:
            $ref: "#/definitions/Message"
        422:
          description: "wrongEntityError is an error object to inform that the provided\
            \ object was not correctly formated"
          schema:
            $ref: "#/definitions/genericErrorModel"
        503:
          description: "databaseError is an error object to tell what is happening\
            \ when we encounter issue with database"
          schema:
            $ref: "#/definitions/genericErrorModel"
        default:
          description: "genericError general error when unexpected errors occured"
          schema:
            $ref: "#/definitions/genericErrorModel"
...
definitions:
  Avatar:
    type: "object"
    required:
    - "link"
    - "name"
    properties:
      id:
        type: "integer"
        format: "uint64"
        description: "id of the avatar"
        minimum: 0.0
        x-go-name: "IDAvatar"
      link:
        type: "string"
        description: "Path into server"
        x-go-name: "Link"
      name:
        type: "string"
        description: "Avatar name"
        x-go-name: "Name"
    title: "Avatar object"
    description: "Avatar object store default avatar images you can use within an\
      \ association.\nRequired apply only for creation of the object. Id is required\
      \ only on update."
    x-go-package: "github.com/titouanfreville/popcubeapi/models"
...
  Folder:
    type: "object"
    required:
    - "id_message"
    - "link"
    - "name"
    - "type"
    properties:
      id:
        type: "integer"
        format: "uint64"
        description: "id of the folder"
        minimum: 0.0
        x-go-name: "IDFolder"
      id_message:
        type: "integer"
        format: "uint64"
        description: "id of the message folder is in"
        x-go-name: "IDMessage"
      link:
        type: "string"
        description: "path to the folder in the server"
        x-go-name: "Link"
      name:
        type: "string"
        description: "folder name"
        x-go-name: "Name"
      type:
        type: "string"
        description: "type if the folder (extension, snippet)"
        x-go-name: "Type"
    title: "Folder object"
    description: "A folder is a file larger than a simple message."
    x-go-package: "github.com/titouanfreville/popcubeapi/models"
...
  Message:
    type: "object"
    required:
    - "date"
    - "id_channel"
    - "id_user"
    properties:
      content:
        type: "string"
        description: "Content of the message"
        x-go-name: "Content"
      date:
        type: "integer"
        format: "int64"
        description: "Date the message was sent at"
        x-go-name: "Date"
      id:
        type: "integer"
        format: "uint64"
        description: "id of the message"
        minimum: 0.0
        x-go-name: "IDMessage"
      id_channel:
        type: "integer"
        format: "uint64"
        description: "Channel reference id"
        x-go-name: "IDChannel"
      id_user:
        type: "integer"
        format: "uint64"
        description: "User reference id"
        x-go-name: "IDUser"
    title: "Message object"
    description: "Message informations and content"
    x-go-package: "github.com/titouanfreville/popcubeapi/models"
...
  deleteMessage:
    type: "object"
    properties:
      Message:
        description: "in:body"
        $ref: "#/definitions/deleteMessageModel"
    title: "deleteMessage return object to confirm correct deletion of an item."
    x-go-package: "github.com/titouanfreville/popcubeapi/api"
  deleteMessageModel:
    type: "object"
    properties:
      message:
        type: "string"
        description: "More information about why is it or isn't it removed"
        x-go-name: "Message"
      removed_object:
        type: "object"
        description: "The object we where trying to remove"
        properties: {}
        x-go-name: "Object"
      status:
        type: "integer"
        format: "int64"
        description: "Status"
        x-go-name: "Status"
      success:
        type: "boolean"
        description: "Correctly removed ?"
        x-go-name: "Success"
    title: "deleteMessageModel is an object to confirm correct deletion of an item."
    x-go-package: "github.com/titouanfreville/popcubeapi/api"
...
  genericErrorModel:
    type: "object"
    properties:
      detailed_error:
        type: "string"
        description: "Detail of error"
        x-go-name: "DetailedError"
      id:
        type: "string"
        description: "ID for the error"
        x-go-name: "ID"
      is_oauth:
        type: "boolean"
        x-go-name: "IsOAuth"
      message:
        type: "string"
        description: "Error's message"
        x-go-name: "Message"
      request_id:
        type: "string"
        description: "Id of the request if exist"
        x-go-name: "RequestID"
      status_code:
        type: "integer"
        format: "int64"
        description: "EgenericErrorModelrror's status code"
        x-go-name: "StatusCode"
    title: "AppError Type used to structure error reporting for popcube chat project."
    x-go-name: "AppError"
    x-go-package: "github.com/titouanfreville/popcubeapi/utils"
...